Avoid using `return` outside of a function
Node.js allows this due to the way globals are implemented there (it wraps
all modules into a function), but it's not a correct JavaScript. One of
the side-effects of this is that bundlers like nexe/nexe don't work with
applications that use eccrypto, as these bundlers use Acorn to parse
all sources into AST and then combine them into a single JS file, and
Acorn chokes on out-of-function return.